python-data-engineer/week7/mini_project/requirements.txt
# Message Queues
paho-mqtt==1.6.1
kafka-python==2.0.2
redis==4.5.1

# Database
psycopg2-binary==2.9.5
SQLAlchemy==1.4.41

# Processing
numpy==1.24.2
pandas==1.5.3
scikit-learn==1.2.2

# Monitoring
prometheus-client==0.16.0
grafana-api==1.0.3

# Utils
python-dotenv==1.0.0
pydantic==1.10.7
structlog==23.1.0
tenacity==8.2.2 